basic skeleton of projectsrc main java with classes
