+++ .gitignore
@@ -2,3 +2,4 @@
 
 Record-Store-DB.mwb
 Record-Store-DB.mwb.bak
+/bin/
