--- QueryLogic.java
@@ -1,20 +1,0 @@
-/**
-* QueryLogic
-*/
-
-
-
-
-public class QueryLogic{
-	
-	private Connection conn;
-	
-	public QueryLogic(Connection c){
-		this.conn = c;
-	}
-	
-	public void getAlbumStockAtAllLocations(String albumTitle){
-		System.out.println("Thinking this will print a list of availability by location");
-	}
-	
-}
