--- UI.java
@@ -1,21 +1,0 @@
-/**
-* UI class
-*/
-
-
-
-
-public class UI{
-	
-	private QueryLogic ql;
-	private DatabaseUpdater updater;
-	
-	public void menu(){
-		System.out.println("Thinking this will be a switch");
-	}
-	
-	public static void main(String[] args){
-		System.out.println("Welcome...");
-		this.menu();
-	}
-}
